Fix image resizing to do the appropriate thing
It will not do aspect fill when necessary, but will also maintain
original aspect ratios when it is needed. It will also never resize up.

It's a bit complex, which is unfortunate, but I'm sure Panda or Nong
can make it better. This kind of math is not my strong suit.